requirements.txt
python-dotenv==1.0.1
openai==1.67.0
requests==2.32.3
uvicorn==0.34.0
fastapi==0.115.4
jinja2==3.1.6
python-multipart==0.0.20
